+++ lib/core/constants/app_images.dart
@@ -4,6 +4,6 @@ class Assets {
   
   /// Assets for imagesLogo 1
   /// assets/images/logo 1.png
-  static const String whatAppLogo = "assets/images/logo 1.png";
+  static const  imagesLogo  = "assets/images/logo 1.png";
 }
