+++ lib/features/splash/presentation/views/splash_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mini_chat_app_mentor_academy/features/splash/presentation/view/widgets/splash_view_body.dart';
+import 'package:mini_chat_app_mentor_academy/features/splash/presentation/views/widgets/splash_view_body.dart';
 
 class SplashView extends StatelessWidget {
   const SplashView({super.key});
