+++ pubspec.yaml
@@ -91,6 +91,6 @@ flutter:
   # see https://flutter.dev/to/font-from-package
 flutter_assets:
   assets_path: assets/images/
-  output_path: lib/core/utils
+  output_path: lib/core/constants
   filename: app_images.dart
   field_prefix: null
